Refactor Makefile as bash scripts
This allows me to write bash easier since some bash stuff is *not*
directly applicable to make without some escaping that I was unable to
do.

A major side effect of this is that I reorganized the whole project so
that there is a `services/` subdirectory for the services that were at
root directory, and a `scripts/` subdirectory.